Changes to query format for facets; backend does only one DB query for all count views in a query.
This sets a more specific data format for facets, ie each must be on a single field even if the backend might rewrite that field. It is the first step toward doing hierarchical facets. This change also improves backend efficiency by doing all count operations on a single DB query (which was possible to do before the query change but not as straightforward).