Incomplete version of new backend using Whoosh; some cleanup and small improvements to programs.
Currently it is possible to create a Whoosh index and run backend queries against it, but not to run the backend as a server.

Cleanup is especially changing the queries.py code to not add fields to the query objects it is given. Improvement is especially more debugging options for querydb.