Option to set required keys for result-paginated views (right now, that is count views) that will be included regardless of pagination.
The result preserves the expected sorting. Required keys that were outside the requested page are prepended or appended (depending their position in the sorted order), while no items are removed. Therefore the number of items in the result may be larger than the page size, up to the page size plus number of required keys. The frontend shows the required values as expected, but since it doesn't scroll to the selected value when it moves in the list the behaviour may still be confusing.