Factored out the index page head section and body header; more general system replacing fixed about page; moved favicon to domain directories.
The head section (added to any fixed lines in the source index file) and body head are specified as HTML fragment files. You can now specify arbitrary HTML files to include with the same replacements (CSS includes, etc.) that the index file gets, so there is no need for a fixed about file in the make system; if you want one you add it as an HTML file and reference it in the body header.