Add Query.save and .load functions
.save will return the list of constraints in a string (created by
JSON.stringify) which can be saved to a file.

.load will take the string produced by .save and generate constraints from it,
adding them into the globalQuery.

These functions can be used for saving multiple queries so you can compare
them later (in future tabs).

The main issue with this approach is that only the name and value attributes of
the queries are saved. All watchers are lost. So this won't be too great for
saving a query and coming back to it later, but it's still good for
comparisons.