Facets should save their information in an array.
This is needed so other views can use selections made in the facets view.

This change is taken from the xkcdplottimelines branch, so it might end up in
a merge conflict, but there's no other way to do it. I can't do a cherry-pick
because the commit that introduced this change also made other modifications
that I don't want.