Fix selection count issue in the cluster view.
Previously, when you selected a box, expanded and then shrunk it back to
original size (without releasing the mouse), the count wouldn't decrease
properly. It wouldn't show show the actual count in the selection, but
somewhere between the biggest size and smallest size of the selection.

This was caused by the fact that the quadtree.visit wasn't visiting all nodes
which were previously in the selection, so some weren't being removed from
constraintIds.

Solution is to add/remove stuff from constraintIds in the same place where the
circles were rendered as selected, and display the counts at the end of
brushed().