Display hover-line at all times when mouse is over SVG.
Mousemove on the lines and mouseover on the SVG were conflicting. Solution was
to create a rect below the lines (which means it has to show up before the
lines at the same depth as the lines in the SVG). This made the hover-line
show up when over the graph, but not when the mouse is over the line. Solution
is to repeat the callbacks for the hover-line in the mouse events of the lines.