Create a new reset-password procedure.
User is now required to input their old password on the user settings page
before being able to set a new one (thus increasing security). This means that
we need a new form for resetting the password through the emailed link.

Admin doesn't need to input the old password when resetting a user's password,
though, so that field should be hidden. A better idea would be to create
separate templates (depending on whether the user or admin is looking at the
page), but that would be overkill if we don't need to add any other forms.