constraints list: Removed cross-module button clicks by DOM lookup.
Having code break modularization like that leads to really unexpected results, so I'm removing it.

As far as I can tell the call to the map clear button wasn't effecting anything. The call to the tsne clear button was making the tsne view clear its brush and selection styles when its constraint was removed, but also when any constraint from any view was removed.