web: Wait for Postgres database to be created
When first running the postgres container, the database will need to be
created, but the web container will run at almost exactly the same time,
so it'll try to connect to a non-existent database and crash. The
solution is to try to connect, and if it fails, wait a bit and try
again. This seems like a hack, but it's not much different than timeout
values on HTTP requests; plus, it avoids having to modify the Postgres
image to let the web container know it's in the process of creating a
DB (which would be tedious to maintain, and hard to debug).