Save notes to localStorage
We're not storing them to the database since that would require extra
changes to deal with tracking which user submitted what (there may be
several users sharing a Google account during our study).

This way, the notes survive a refresh, which is the only amount of
saving that we need right now.