Allow normal text search in Facets view
Each search term is prefixed with the facet's name so the search term
ends up being something like "location:germany". This doesn't work for
all facets, but that should be fixed in the backend, not in the
frontend.

Also, this implementation depends on the facet name being the correct
prefix. If we want to rename the facet name, we'll have to come up with
a different way of storing the prefix.